requirements.txt
streamlit
tensorflow
opencv-python
numpy
scikit-learn
Pillow
matplotlib
